♻️ (refactor): UserManagementView to child components
